fix: waring for comparison between int & size_t

+++ src/sort_insertion.c
@@ -5,7 +5,7 @@
 #include "sort_insertion.h"
 
 static void swap(void *arr, size_t width, int i, int j);
-static bool isSorted(void *arr, size_t width, int n, ComparatorFn cmp);
+static bool isSorted(void *arr, size_t width, size_t n, ComparatorFn cmp);
 
 void insertionSort(void *arr, size_t width, int n, ComparatorFn cmp) {
   for (int i = 1; i < n; i++) {
@@ -27,7 +27,7 @@ static void swap(void *arr, size_t width, int i, int j) {
   memcpy(arr + j * width, temp, width);
 }
 
-static bool isSorted(void *arr, size_t width, int n, ComparatorFn cmp) {
+static bool isSorted(void *arr, size_t width, size_t n, ComparatorFn cmp) {
   for (size_t i = 1; i < n; i++) {
     if (cmp(arr + (i - 1) * width, arr + i * width) > 0) {
       return false;
