+++ src/sort_merge_bottom_up.c
@@ -7,7 +7,7 @@
 
 #define MIN(a, b) (a < b ? a : b)
 
-static bool isSorted(void *arr, size_t width, int n, ComparatorFn cmp);
+static bool isSorted(void *arr, size_t width, size_t n, ComparatorFn cmp);
 
 
 void bottomUpMergeSort(void *arr, size_t width, int n, ComparatorFn cmp) {
@@ -29,7 +29,7 @@ void bottomUpMergeSort(void *arr, size_t width, int n, ComparatorFn cmp) {
 }
 
 
-static bool isSorted(void *arr, size_t width, int n, ComparatorFn cmp) {
+static bool isSorted(void *arr, size_t width, size_t n, ComparatorFn cmp) {
   for (size_t i = 1; i < n; i++) {
     if (cmp(arr + (i - 1) * width, arr + i * width) > 0) {
       return false;
