+++ src/sort_merge_top_down.c
@@ -7,7 +7,7 @@
 
 #define CUT_OFF 6
 
-static bool isSorted(void *arr, size_t width, int n, ComparatorFn cmp);
+static bool isSorted(void *arr, size_t width, size_t n, ComparatorFn cmp);
 
 void merge(
     void *arr,
@@ -73,7 +73,7 @@ void topDownMergeSort(void *arr, size_t width, int n, ComparatorFn cmp) {
 }
 
 
-static bool isSorted(void *arr, size_t width, int n, ComparatorFn cmp) {
+static bool isSorted(void *arr, size_t width, size_t n, ComparatorFn cmp) {
   for (size_t i = 1; i < n; i++) {
     if (cmp(arr + (i - 1) * width, arr + i * width) > 0) {
       return false;
